[DEBUG] Updating debug scripts to compare two directories
This is a major refactor on the diff.py script. Now it supports
comparing two directories of pass dumps on the same passes as each
other. If the pass order is different or there are passes in between,
the script will skip until the same pass is found. It also ignores
common cleanups to avoid confusion.

The script is still part of the debug_all_passes, but can also be used
on its own on existing directories. We can now chose which is the temp
directory that the script will dump, so that it's easier to compare with
a baseline.